Use directly official libnbd go bindings (#2081)
* Use official libnbd go bindings

Currently, CDI relies for the libnbd bindings on:
https://github.com/mrnold/go-libnbd

This PR proposes a way to directly uses the official bindings from:
https://gitlab.com/nbdkit/libnbd/-/tree/master/golang

The major issue in using the official go bindings and bazel consists
that gazelle doesn't support pkg-conf. Consequently, the link and cflags
are not automatically generated together with the BUILD.bazel file.
A possible workaround is to add the missing flags using buildozer.

This approach avoids the needs of an extra repository to fix the copt
and linkopt.

Signed-off-by: Alice Frosi <afrosi@redhat.com>

* Update vendor and importer BUILD file

Signed-off-by: Alice Frosi <afrosi@redhat.com>

* fetch newer libnbd version

Signed-off-by: Alice Frosi <afrosi@redhat.com>