Fix VerifySparse to look at content size and not size on disk (#2168)
We consistently fail on the check that `info.VirtualSize >= info.ActualSize`
on a real external nfs sever with different block size to what we're using upstream.

This happens because qemu-img info gives us the size on disk in ActualSize.
For the purpose of sparseness, I believe we are interested in the content size and not the padded size on disk.

Signed-off-by: Alex Kalenyuk <akalenyu@redhat.com>