Build builder for multiple platforms
Requires buildah and qemu-user-static to be installed locally

Test for qemu-user-static: the failure mode for when it doesn't
exist is surprising. Tolerate it not existing on non-Linux, for
example on macOS.

Get a bazel for the current architecture.

To help native builds on aarch64 (which we need to call arm64 for
the download link)

Signed-off-by: Maya Rashish <mrashish@redhat.com>