[release-v1.43] Delete erroneous DVs on DataImportCron digest update (#2220)
Manual backport of #2169

When a DataImportCron import DV is condition Running=False with
Reason=Error it indicates this DIC might get stuck with this DV forever,
so no new import DVs will be created even if the source sha256 is
updated. With this change, when digest is updated, before creating the
new DV, we simply delete the erroneous DV if necessary.

Added a timeout to the importer cmd wait for server ready file, so it
exits with error and termination message used by DV Running condition.

Also includes some DataImportCron tests improvements and cleanup.

Signed-off-by: Arnon Gilboa <agilboa@redhat.com>