Run update-ca-trust whenever running cdi-source-update-poller
Do this using a shell wrapper that runs it first, then calls the
actual executable with the same argument.

Background:
Our switch to building CentOS Stream images from RPMs has introduced
a small problem that root CAs are not installed.

As a temporary measure, we are running update-ca-trust, a very quick
running binary.

We are working around it using BUILD.bazel to avoid making breaking
changes to non-bazel builds.

Signed-off-by: Maya Rashish <mrashish@redhat.com>