Allow running most makefile targets without docker
Uses podman and buildah - determined by ${CDI_CRI} and
${CDI_CONTAINER_BUILDCMD}. These are just 'docker' if docker is
chosen.

Somewhat confusingly, prefers building with buildah if it exists,
as that's the scenario for the CI and we want to use buildah to
build a multi-arch container.

Uses KUBEVIRTCI_RUNTIME as an override tunable, as that applies to
kubevirtci code as well.

Invert logic testing KUBEVIRTCI to be podman to CDI_CRI not to be
docker -- this is because the CDI_CRI in the podman case includes
some extra parameters.

Signed-off-by: Maya Rashish <mrashish@redhat.com>