Build builder for multiple platforms
If buildah is used for building containers and not docker,
reqires qemu-user-static. We have an explicit test for it
being installed as the failure mode is somewhat confusing.

If buildah is not available, keep using docker to build the
builder, and only build for the native arch (no additional
requirements needed)

Get bazel for the current architecture.

To help native builds on aarch64 (which we need to call arm64 for
the download link)

Signed-off-by: Maya Rashish <mrashish@redhat.com>