[release-v1.43] Fix smart clone request size update (#2289)
* Fix smart clone request size update

In case the pvc got an actual size as the expected
size there wasn't an update of the pvc spec with the
actual user request size, which left the pvc spec with
a smaller size then the user requested(the data size).
This caused a discrepancy when trying to restore such pvc,
which restored a pvc with the small size instead of the user request
size.

Signed-off-by: Shelly Kagan <skagan@redhat.com>

* review fixes

Signed-off-by: Shelly Kagan <skagan@redhat.com>

* cherry pick fix

Signed-off-by: Shelly Kagan <skagan@redhat.com>

* Fix smart clone update request size condition

Signed-off-by: Shelly Kagan <skagan@redhat.com>