Small shim so libvddk-test-plugin.so can run on older glibc (#2297)
vddk-test is special, it creates a dynamic library that is
loaded by the importer for test suite runs.

When running tests on a cluster built with an importer built from a
different base image, e.g. when RedHat builds with a CentOS8
equivalent, we might be attempting to load libvddk-test-plugin.so on
the CentOS8 importer. This breaks, as it doesn't have newer symbols.

There's only one new glibc symbol used, for fstat. Let's use
the older glibc definition of fstat, __fxstat explicitly.

Signed-off-by: Maya Rashish <mrashish@redhat.com>