[release-v1.38] Start smart clone controller from datavolume controller when needed #2265 (#2307)
* Start smart clone controller from datavolume controller when needed (#2265)

* periodic sync CSI snapshot CRD check

It was possible for the CSI snapshot CRD check to fail silently and
prevent the smart clone controller from starting during the cdi deployment
pod start up. This would prevent smart clone from working properly.

This adds a periodic sync of 1 minute for checking the CRDs. We also
log failures that are not is not found so we can more easily detect this
situation as humans.

Signed-off-by: Alexander Wels <awels@redhat.com>

* Change location of the start controller call.

Signed-off-by: Alexander Wels <awels@redhat.com>

* Modifications needed for backport. In particular added check for both betav1
and v1 crds.

Signed-off-by: Alexander Wels <awels@redhat.com>