Allow running all makefile targets without docker installed, using podman/buildah (#2279)
* Remove unnecessary difference between docker and podman

docker accepts the same command line podman required.

Signed-off-by: Maya Rashish <mrashish@redhat.com>

* Update openshift ci dockerfile

Update base image so we can build our code - requires newer Go

Signed-off-by: Maya Rashish <mrashish@redhat.com>

* Specify a tag to be built when building openshift ci images

Signed-off-by: Maya Rashish <mrashish@redhat.com>

* Allow running most makefile targets without docker

Uses podman and buildah - determined by ${CDI_CRI} and
${CDI_CONTAINER_BUILDCMD}. These are just 'docker' if docker is
chosen.

Somewhat confusingly, prefers building with buildah if it exists,
as that's the scenario for the CI and we want to use buildah to
build a multi-arch container.

Uses KUBEVIRTCI_RUNTIME as an override tunable, as that applies to
kubevirtci code as well.

Invert logic testing KUBEVIRTCI to be podman to CDI_CRI not to be
docker -- this is because the CDI_CRI in the podman case includes
some extra parameters.

Signed-off-by: Maya Rashish <mrashish@redhat.com>

* Build builder for multiple platforms

If buildah is used for building containers and not docker,
reqires qemu-user-static. We have an explicit test for it
being installed as the failure mode is somewhat confusing.

If buildah is not available, keep using docker to build the
builder, and only build for the native arch (no additional
requirements needed)

Get bazel for the current architecture.

To help native builds on aarch64 (which we need to call arm64 for
the download link)

Signed-off-by: Maya Rashish <mrashish@redhat.com>

* Combine conditionals, combine 2 env variables

Make buildah & podman come together, eliminate use of `which`

Signed-off-by: Maya Rashish <mrashish@redhat.com>

* For clarity, note output-arch is meant for bazel architecture

Signed-off-by: Maya Rashish <mrashish@redhat.com>

* Remove unnecessary diff between docker and podman

Seems to work anyway.

Signed-off-by: Maya Rashish <mrashish@redhat.com>

* Warn instead of failing when qemu-user-static isn't installed

Signed-off-by: Maya Rashish <mrashish@redhat.com>