Allow builder with unconfined security profile if env variable is set (#2330)
Running our builder on fedora 34 will fail due to a new syscall in centos
that is not support by the security profile in fedora 34. Adding the unconfined
security profile to the builder fixes this. But we don't want to enable this by
default. So check for environment variable before setting it.

export DISABLE_SECCOMP="--security-opt seccomp=unconfined"

Signed-off-by: Alexander Wels <awels@redhat.com>