Fix race condition in upload-controller to keep PVC annotations after a clone error (#2363)
* Modify upload-controller to keep the annotations of a PVC if a pod fails in another controller

When a pod fails, some annotations (running condition, pod phase...) are updated in the affected PVC to let other controllers know.

However, due to a lack of synchronization between some controllers, a race condition could happen where, if a pod succeeds just after a pod fails in a different controller, the annotations set in the error handling would just be updated inapropiately.

This commit modifies the upload-controller to check for clone failures before updating annotations.

Signed-off-by: Alvaro Romero <alromero@redhat.com>

* Add unit tests to updateUploadAnnotations

Signed-off-by: Alvaro Romero <alromero@redhat.com>

* Update functional tests in cloner_test to check for conditions and annotations after a pod fails

Signed-off-by: Alvaro Romero <alromero@redhat.com>

* Improve error handling in source-clone pod to avoid overwritting the DV running condition

Signed-off-by: Alvaro Romero <alromero@redhat.com>