VDDK: look for disk in root snapshot parent.
If the target disk name wasn't found on the current hardware config or
in the snapshot list, look in the parent backing file of the root
snapshot. Sometimes base file names (like [dc] test/test.vmdk) are
absent from the snapshot tree, for example if the VM is marked
"consolidation required" in VMware.

Signed-off-by: Matthew Arnold <marnold@redhat.com>